Hackathon_B/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.6.1
requests==2.32.3
python-dotenv==1.0.1
openai==1.58.1
pinecone-client==3.2.2
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.2
PyPDF2==3.0.1
python-docx==1.1.2
langchain==0.3.0
langchain-openai==0.3.0
sentence-transformers==2.7.0
tiktoken==0.7.0
aiofiles==24.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
